move non-check funcs to different file
