+++ pkg/collector/corechecks/windows_event_log/check_windows.go
@@ -168,31 +168,6 @@ func (c *Check) bookmarkPersistentCacheKey() string {
 	return fmt.Sprintf("%s_%s", c.ID(), "bookmark")
 }
 
-func alertTypeFromLevel(level uint64) (agentEvent.EventAlertType, error) {
-	// https://docs.microsoft.com/en-us/windows/win32/wes/eventmanifestschema-leveltype-complextype#remarks
-	// https://learn.microsoft.com/en-us/windows/win32/wes/eventmanifestschema-eventdefinitiontype-complextype#attributes
-	// > If you do not specify a level, the event descriptor will contain a zero for level.
-	var alertType string
-	switch level {
-	case 0:
-		alertType = "info"
-	case 1:
-		alertType = "error"
-	case 2:
-		alertType = "error"
-	case 3:
-		alertType = "warning"
-	case 4:
-		alertType = "info"
-	case 5:
-		alertType = "info"
-	default:
-		return agentEvent.EventAlertTypeInfo, fmt.Errorf("invalid event level: '%d'", level)
-	}
-
-	return agentEvent.GetAlertTypeFromString(alertType)
-}
-
 func (c *Check) renderEventValues(winevent *evtapi.EventRecord, ddevent *agentEvent.Event) error {
 	// Render the values
 	vals, err := c.evtapi.EvtRenderEventValues(c.systemRenderContext, winevent.EventRecordHandle)
@@ -420,18 +395,6 @@ func (c *Check) Configure(senderManager sender.SenderManager, integrationConfigD
 	}
 
 	return nil
-}
-
-func compileRegexPatterns(patterns []string) ([]*regexp.Regexp, error) {
-	var err error
-	res := make([]*regexp.Regexp, len(patterns))
-	for i, pattern := range patterns {
-		res[i], err = regexp.Compile(pattern)
-		if err != nil {
-			return nil, fmt.Errorf("error compiling regex pattern '%s': %v", pattern, err)
-		}
-	}
-	return res, nil
 }
 
 func (c *Check) validateConfig() error {
